# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Collab API",
    "description": "API service for projects",
    "version": "1.0.0",
    "contact": { "email": "brbryant2639@gmail.com" }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://mywebsite.com",
      "description": "Live address for api resource"
    }
  ],
  "host": "mywebsite.com",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
...
  "paths": {
...
    "/watchlist": {
      "get": {
        "tags": ["watchlist"],
        "summary": "Get a list of all accounts",
        "description": "Restricted to admin users.",
        "operationId": "getAllAccounts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "An array of all accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "5eb12e197e06a76ccdefc121"
                      },
                      "firstName": {
                        "type": "string",
                        "example": "Joe"
                      },
                      "lastName": {
                        "type": "string",
                        "example": "Smith"
                      },
                      "email": {
                        "type": "string",
                        "example": "joe@mail.com"
                      },
                      "role": {
                        "type": "string",
                        "example": "User"
                      },
                      "created": {
                        "type": "string",
                        "example": "2020-05-05T09:12:57.848Z"
                      },
                      "updated": {
                        "type": "string",
                        "example": "2020-05-08T03:11:21.553Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "tags": ["watchlist"],
        "summary": "Create a watchlist item",
        "description": "Restricted to admin users.",
        "operationId": "createAccount",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Joe"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Smith"
                  },
                  "email": {
                    "type": "string",
                    "example": "joe@mail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "some_password!"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "some_password!"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["Admin", "User"]
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password",
                  "confirmPassword",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully, verification is not required for users created with this endpoint. The details of the new user are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "5eb12e197e06a76ccdefc121"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Joe"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Smith"
                    },
                    "email": {
                      "type": "string",
                      "example": "joe@smith.com"
                    },
                    "role": {
                      "type": "string",
                      "example": "User"
                    },
                    "created": {
                      "type": "string",
                      "example": "2020-05-05T09:12:57.848Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email is already registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email 'joe@mail.com' is already registered"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid, or the user does not have access to perform the action",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
    }
  }
}